Complete refactoring: update tests and remove redundant files
- Update test imports to use the new module structure
- Remove old workflow files (replaced by unified SpellingGrammarWorkflow)
- Remove files moved to new architecture layers:
  - spellingGrammarWorkflow.ts → workflows/SpellingGrammarWorkflow.ts
  - parallelSpellingGrammarWorkflow.ts → merged into unified workflow
  - postProcessing.ts → application/processing.ts
  - workflowHelpers.ts → distributed across application layer
  - documentProcessor.ts → infrastructure/documentProcessor.ts
  - validation.ts → unused (validation now in domain objects)
  - errors.ts → domain/errors.ts

All tests pass and TypeScript compilation succeeds.
